Add a CPU check and CI
Similar to https://github.com/AI-Hypercomputer/torchprime/pull/13, we
also add a CPU github action. This action will run `pytest` on the repo.

Currently there is only one test, which is the Llama test in
torch_xla_models.

In order to run the test today, we need a HF_TOKEN. I created a personal
read only token and
https://github.com/AI-Hypercomputer/torchprime/issues/14 tracks avoiding
the need for HF_TOKEN, after which I'll need to remember to invalidate
the token.