Improve xpk workflows
Tick off some of https://github.com/AI-Hypercomputer/torchprime/issues/33
with this change. We add a `tp` command that will be automatically placed
in the user's $PATH when installing the repo. It has two verbs:

- tp use [...]: sets the XPK cluster/project/zone and profile/xla dump
  GCS directory once for all future commands.
- tp run [...]: runs a python script on the XPK cluster

In the future, we could add utilities anchoring off the `tp` CLI. For
example, we could add a `tp log` that follows the log of a workload, or
a `tp bisect` that determines the software version corresponding to a
regression.