[docker] Only run `pip install -e .` when changing pyproject.toml
This is an optimization. Previously, if I change any file in the docker
context (e.g. any python file), then all packages are reinstalled,
resulting in a big container layer that's slow to push and pull.

Now if I change a python file, that file will simply be copied into the
last docker layer. No more `pip install -e .` thanks to the docker
cache.

Addresses a point in
https://github.com/AI-Hypercomputer/torchprime/issues/33.