Use different docker URLs so everyone's setup don't collide
Instead of uploading to tpu-pytorch/llama, we upload to a docker image
name that is scoped under the current username, and we also upload to a
"datetime + random string" tag. This way everyone can run `tp run` and
their images won't overwrite one another.

Also adds a `tp use --docker-url` option to override the GCP project the
docker image is uploaded to. But this is optional. Usually the XPK
cluster's project will allow everyone to push to it, but we could also
override to push to e.g. tpu-pytorch if we want.