Fix profiling of torchax models
Since torch_xla uses hydra, the CLI for setting profile dir changed
slightly. In order to profile torchax training, we refactor both to
use hydra and the same config name for setting profile dir.

I didn't do that in this PR, but it will be simpler to extend the
torchax trainer to train both llama and mixtral via configs later.

Tested:

  python3 torchprime/experimental/torchax_models/run.py global_batch_size=16

  tp run torchprime/experimental/torchax_models/run.py \
    global_batch_size=256 \
    model_type=405B \
    seqlen=8192 \
    use_custom_offload=True \
    use_custom_mesh=True \
    model_impl=scan \
    tp=4 \
    unroll_layers=1