Start of a generic config based model sharder
Motivation

In Hugging Face transformer models, we have been using FSDPv2 [1] and
hardcoded mark_sharding calls to annotate tensors with sharding
constraints. This is not very extensible. FSDPv2 doesn't support 2D
sharding, and changing sharding config requires modifying the model
code.

This PR introduces the start of a config based model sharder that
generalizes to arbitrary sharding schemes supported by GSPMD, e.g. 2D
FSDP+TP sharding, expert parallelism, etc.

Design

The user can specify the names of weights to shard and their partition
spec. They can also shard certain activations by specifying the names of
the `nn.Module` that outputs that activation:

  ```yaml
  # Weights
  model.embed_tokens.weight: ['fsdp', null]
  model.layers.*.self_attn.q_proj.weight: ['fsdp', null]
  model.layers.*.self_attn.k_proj.weight: [null, 'fsdp']
  model.layers.*.self_attn.v_proj.weight: [null, 'fsdp']
  model.layers.*.self_attn.o_proj.weight: ['fsdp', null]
  model.layers.*.mlp.gate_proj.weight: ['fsdp', null]
  model.layers.*.mlp.up_proj.weight: ['fsdp', null]
  model.layers.*.mlp.down_proj.weight: [null, 'fsdp']
  model.layers.*.input_layernorm.weight: ['fsdp']
  model.layers.*.post_attention_layernorm.weight: ['fsdp']
  model.norm.weight: ['fsdp']
  lm_head.weight: ['fsdp', null]

  # Activations
  model.embed_tokens: ['fsdp', null, null]
  model.layers.*.self_attn: ['fsdp', null, null]
  model.layers.*: ['fsdp', null, null]
  model.lm_head: ['fsdp', null, null]
  ```

In this PR, they can shard a model by passing this config to
`shard_model_from_config` in Python. Subsequently we'll support sharding
a model from a yaml configuration file as well.

Under this scheme, it's immediately obvious how is a particular param or
activation sharded. We can later create different configs for FSDP, 2D
sharding, and so on.

To shard parameters, we traverse the state dict and call mark_sharding.

To shard activations, we replace the matching `nn.Module` with a
`ShardedModule` wrapper that runs the underlying module, and then call
`mark_sharding` to shard the output.

The `mark_sharding` is a pluggable callback so that we could use this
with both `torch_xla` and `torchax`.

[1]: https://github.com/pytorch/xla/issues/6379